Support splitting hostname across multiple lines.
On windows, each line in the hosts file can only contain a
maximum of 10 hosts.

This add the logic to limit the number of host mapped to a single
line to a maximum number.

By default this is set to no limit (infinite).
When Windows is detected, the limit is automatically set to 10.

A different value can be configured by setting the
`KUBEFWD_MAX_HOSTNAMES_PER_LINE` environment variable to the
maximum number of hosts per line.

This is useful because IMHO, having multiple mappings to a single line
quickly become messy.
So a user can decide to set this to 1.

.

.